rename customLogging to custom_logging to conform to PEP8 standards
